restructure health service to align with other gs services
