+++ service/health/key/key.go
@@ -4,39 +4,26 @@ package key
 type Health string
 
 const (
-	// Red represents an unhealthy cluster
-	Red Health = "red"
-	// Yellow represents a passable cluster
-	Yellow Health = "yellow"
-	// Green represents a healthy cluster
-	Green Health = "green"
-	// Default is used when a condition cannot otherwise be determined
-	Default Health = Green
+	// HealthRed represents an unhealthy cluster
+	HealthRed Health = "red"
+	// HealthYellow represents a passable cluster
+	HealthYellow Health = "yellow"
+	// HealthGreen represents a healthy cluster
+	HealthGreen Health = "green"
+	// HealthDefault is used when a condition cannot otherwise be determined
+	HealthDefault Health = HealthGreen
 )
 
 // LifecycleState is an enumeration of states correspoding to certain cluster conditions
 type LifecycleState string
 
 const (
-	// Normal cluster state represents a cluster which is not creating, upgrading, or deleting
-	Normal LifecycleState = "normal"
-	// Creating cluster state represents a cluster which is still being created
-	Creating LifecycleState = "creating"
-	// Upgrading cluster state represents a cluster which is still being upgraded
-	Upgrading LifecycleState = "upgrading"
-	// Deleting cluster state represents a cluster which is still being deleting
-	Deleting LifecycleState = "deleting"
+	// StateNormal cluster state represents a cluster which is not creating, upgrading, or deleting
+	StateNormal LifecycleState = "normal"
+	// StateCreating cluster state represents a cluster which is still being created
+	StateCreating LifecycleState = "creating"
+	// StateUpgrading cluster state represents a cluster which is still being upgraded
+	StateUpgrading LifecycleState = "upgrading"
+	// StateDeleting cluster state represents a cluster which is still being deleting
+	StateDeleting LifecycleState = "deleting"
 )
-
-// Providers is an array of providers which are currently implemented for health checking
-var Providers = [...]string{"aws", "azure", "kvm"}
-
-// IsKnownProvider is a helper function returning true iff the given value is in Providers
-func IsKnownProvider(value string) bool {
-	for i := range Providers {
-		if Providers[i] == value {
-			return true
-		}
-	}
-	return false
-}
