+++ service/health/response.go
@@ -1,53 +1,7 @@
 package health
 
-import "github.com/giantswarm/health-service/service/health/key"
-
-// Response holds the data returned from the health searcher endpoint.
+// Response holds the data returned from the health service.
 type Response struct {
 	Cluster ClusterStatus `json:"cluster"`
 	Nodes   []NodeStatus  `json:"nodes"`
-}
-
-// ClusterStatus holds data about the overall health/status of a cluster.
-type ClusterStatus struct {
-	Health key.Health         `json:"health"`
-	State  key.LifecycleState `json:"state"`
-}
-
-// NodeStatusComputeResources holds data about available or requested compute resources.
-type NodeStatusComputeResources struct {
-	CPU         int64 `json:"cpu"`
-	MemoryBytes int64 `json:"memory_bytes"`
-}
-
-// NodeStatusMachineResources holds data about the status of storage on a cluster node.
-type NodeStatusMachineResources struct {
-	CPUCount               int64 `json:"cpu_count"`
-	MemoryCapacityBytes    int64 `json:"memory_capacity_bytes"`
-	MemoryAllocatableBytes int64 `json:"memory_allocatable_bytes"`
-	EphemeralStorageCap    int64 `json:"ephemeral_storage_capacity_bytes"`
-	EphemeralStorageAvail  int64 `json:"ephemeral_storage_allocatable_bytes"`
-}
-
-// NodeStatusIdentity holds data about the identity of a node which is generally static.
-type NodeStatusIdentity struct {
-	Name               string `json:"name"`
-	KubeletVersion     string `json:"kubelet_version"`
-	OperatorVersion    string `json:"operator_version"`
-	Role               string `json:"role"`
-	IP                 string `json:"ip"`
-	Hostname           string `json:"hostname"`
-	InstanceType       string `json:"instance_type"`
-	AvailabilityRegion string `json:"availability_region"`
-	AvailabilityZone   string `json:"availability_zone"`
-}
-
-// NodeStatus represents the health status of a single node in a cluster.
-type NodeStatus struct {
-	Ready            bool                       `json:"ready"`
-	Health           key.Health                 `json:"health"`
-	Identity         NodeStatusIdentity         `json:"identity"`
-	MachineResources NodeStatusMachineResources `json:"machine"`
-	LimitTotals      NodeStatusComputeResources `json:"limit_totals"`
-	RequestTotals    NodeStatusComputeResources `json:"request_totals"`
 }
